src/requirements.txt
opencv-python
numpy
pandas
pillow
matplotlib
ultralytics